# docs/migration/contract/openapi-legacy-baseline.yaml
openapi: 3.1.0
info:
  title: Gia Pha Legacy Backend HTTP Contract (Frozen Baseline)
  version: 1.0.0-legacy-baseline
  description: |
    Frozen contract of every Next.js App Router `/api/**` handler prior to the
    Spring Boot migration (Task 1 of spring-boot-backend-migration).

    Conventions frozen from the legacy implementation:
    * **Error envelope** — every error response is `{"ok": false, "error": {"code", "message", "details?"}}`.
    * **Mixed success envelopes** — most successful responses return the raw
      resource JSON (no envelope). The exceptions that wrap in
      `{"ok": true, "data": ...}` are: `POST /api/auth/register`,
      `PATCH /api/trees/{treeId}/memberships/{userId}`,
      `POST /api/media/upload`, and share-link management responses.
    * **Timestamps** — ISO-8601 UTC strings with milliseconds (`.toISOString()` form).
    * **Authentication** — NextAuth JWT session cookie
      (`next-auth.session-token` / `__Secure-next-auth.session-token`).
      Unauthenticated requests receive `401 UNAUTHORIZED`.
    * **RBAC** — tree-scoped roles ADMIN/EDITOR/VIEWER, owner is immutable
      effective ADMIN. Missing tree → `404 TREE_NOT_FOUND`, missing
      permission → `403 FORBIDDEN`.
    * **Tree discovery** — detail routes for members/relationships/events/media/
      albums accept optional `?treeId=`; without it the legacy handler scans
      every tree to locate the resource. Compatibility_API must preserve this.
...
servers:
  - url: /
...
paths:
...
  /api/trees/{treeId}/relationships:
    parameters: [ { $ref: '#/components/parameters/treeId' } ]
    get:
      tags: [relationships]
      operationId: listRelationships
      description: Requires READ. `?memberId=` returns RelationshipView list for that member.
      parameters:
        - { name: memberId, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Raw array (Relationship[] or RelationshipView[] when memberId given).
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Relationship' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [relationships]
      operationId: createRelationship
      description: |
        Requires CREATE. Symmetric types (SPOUSE/SIBLING/CUSTOM) are stored with
        canonical endpoint order (source <= target). Duplicates by logical key
        rejected 409 DUPLICATE; PARENT_CHILD cycles rejected 409 CYCLE_DETECTED.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateRelationshipInput' } } } }
      responses:
        '201': { description: Raw created relationship (normalized)., content: { application/json: { schema: { $ref: '#/components/schemas/Relationship' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: DUPLICATE or CYCLE_DETECTED, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  /api/relationships/{relationshipId}:
    parameters:
      - { name: relationshipId, in: path, required: true, schema: { type: string } }
      - { name: treeId, in: query, required: false, schema: { type: string } }
    get:
      tags: [relationships]
      operationId: getRelationship
      responses:
        '200': { description: Raw relationship., content: { application/json: { schema: { $ref: '#/components/schemas/Relationship' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [relationships]
      operationId: updateRelationship
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateRelationshipInput' } } } }
      responses:
        '200': { description: Raw updated relationship., content: { application/json: { schema: { $ref: '#/components/schemas/Relationship' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: DUPLICATE or CYCLE_DETECTED, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    delete:
      tags: [relationships]
      operationId: deleteRelationship
      responses:
        '204': { description: Deleted. Empty body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  parameters:
    treeId:
      name: treeId
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: Missing/invalid session (code UNAUTHORIZED).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Forbidden:
      description: Authenticated but lacking the tree permission (code FORBIDDEN).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Resource or tree not found (code NOT_FOUND / TREE_NOT_FOUND).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    ValidationError:
      description: Invalid input (code VALIDATION_ERROR / INVALID_INPUT / INVALID_FILTER; details carries Zod flatten()).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
    RelationshipType: { type: string, enum: [PARENT_CHILD, SPOUSE, SIBLING, ADOPTED, CUSTOM] }
...
    Relationship:
      type: object
      required: [id, treeId, sourceMemberId, targetMemberId, type, createdAt, updatedAt]
      properties:
        id: { type: string }
        treeId: { type: string }
        sourceMemberId: { type: string }
        targetMemberId: { type: string }
        type: { $ref: '#/components/schemas/RelationshipType' }
        customType: { type: string, maxLength: 100, description: Required when type=CUSTOM }
        startDate: { type: string }
        endDate: { type: string }
        notes: { type: string, maxLength: 2000 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateRelationshipInput:
      type: object
      required: [sourceMemberId, targetMemberId, type]
      description: source != target; CUSTOM requires customType.
    UpdateRelationshipInput:
      type: object
      description: Partial; merged value re-validated including duplicate/cycle checks.
...
security:
  - nextAuthSession: []